#перенос команды rebase в cli
