--- bpmdcli/bpmdcli.cmd
@@ -1,1 +1,0 @@
-dotnet %~dp0bpmdcli.dll %*
